Remove SourceInterface
