+++ src/Block/BlockInterface.php
@@ -5,10 +5,11 @@ declare(strict_types=1);
 namespace webignition\BasilCompilationSource\Block;
 
 use webignition\BasilCompilationSource\Line\LineInterface;
-use webignition\BasilCompilationSource\SourceInterface;
+use webignition\BasilCompilationSource\Metadata\MetadataInterface;
 
-interface BlockInterface extends SourceInterface
+interface BlockInterface
 {
+    public function getMetadata(): MetadataInterface;
     public function addLine(LineInterface $statement);
     public function addLinesFromBlock(BlockInterface $block);
