+++ tests/Unit/Line/CommentTest.php
@@ -15,7 +15,6 @@ class CommentTest extends \PHPUnit\Framework\TestCase
         $comment = new Comment($content);
 
         $this->assertSame($content, $comment->getContent());
-        $this->assertSame([$comment], $comment->getSources());
     }
 
     public function testToString()
