+++ tests/Unit/Line/EmptyLineTest.php
@@ -14,7 +14,6 @@ class EmptyLineTest extends \PHPUnit\Framework\TestCase
         $emptyLine = new EmptyLine();
 
         $this->assertSame('', $emptyLine->getContent());
-        $this->assertSame([$emptyLine], $emptyLine->getSources());
     }
 
     public function testToString()
